♻️ extract grid component
