+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-blocks/DatabaseBlock/DatabaseGrid.tsx
@@ -1,0 +1,75 @@
+import { ColDef } from 'ag-grid-community';
+import { AgGridReact } from 'ag-grid-react';
+import { useRef, useState } from 'react';
+
+import { Box } from '@/components';
+
+import { AddButtonComponent } from './AddColumnButton';
+
+export const DatabaseGrid = () => {
+  const gridRef = useRef(null);
+  const [addColumnToggleOpen, setAddColumnToggleOpen] = useState(false);
+
+  const [rowData, setRowData] = useState([
+    { make: 'Tesla', model: 'Model Y', price: 64950, electric: true },
+    { make: 'Ford', model: 'F-Series', price: 33850, electric: false },
+    { make: 'Toyota', model: 'Corolla', price: 29600, electric: false },
+  ]);
+
+  // Column Definitions: Defines the columns to be displayed.
+  const [colDefs, setColDefs] = useState<ColDef[]>([
+    { field: 'make', sort: 'desc' },
+    {
+      field: 'model',
+    },
+    { field: 'price' },
+    { field: 'electric' },
+    {
+      headerComponentParams: {
+        innerHeaderComponent: () =>
+          AddButtonComponent({
+            addColumn,
+            isOpen: addColumnToggleOpen,
+            setIsOpen: setAddColumnToggleOpen,
+          }),
+      },
+      unSortIcon: false,
+      editable: false,
+      sortable: false,
+      spanRows: true,
+      filter: false,
+    },
+  ]);
+
+  const defaultColDef = {
+    flex: 1,
+    filter: true,
+    editable: true,
+    unSortIcon: true,
+  };
+
+  const addColumn = (columnName: string) => {
+    const newColDef: ColDef = {
+      field: columnName,
+    };
+    const addColumn = colDefs.pop();
+    setColDefs((prev) => [...prev, newColDef]);
+    if (addColumn !== undefined) {
+      setColDefs((prev) => [...prev, addColumn]);
+    }
+    setAddColumnToggleOpen(false);
+  };
+
+  return (
+    <Box style={{ height: '100%', width: '100%' }}>
+      <AgGridReact
+        ref={gridRef}
+        rowData={rowData}
+        columnDefs={colDefs}
+        defaultColDef={defaultColDef}
+        domLayout="autoHeight"
+        enableCellSpan={true}
+      />
+    </Box>
+  );
+};
